split routes on server per procedure
