+++ code/mobile/.dockerignore
@@ -1,2 +1,4 @@
 dist/
 .env
+node_modules
+update_shared.sh
