+++ code/mobile/components/ui/select.tsx
@@ -1,3 +1,5 @@
+//@ts-nocheck
+
 import * as SelectPrimitive from "@rn-primitives/select";
 import * as React from "react";
 import { Platform, StyleSheet, View } from "react-native";
