--- code/mobile/lib/food-filtering-options.ts
@@ -1,2 +1,0 @@
-export const FOOD_FILTERING_OPTIONS = ["Often", "Favorite", "Last"] as const;
-export type FoodFilteringOption = (typeof FOOD_FILTERING_OPTIONS)[number];
