+++ code/mobile/tsconfig.json
@@ -6,11 +6,10 @@
     "jsx": "react-jsx",
     "paths": {
       "@/*": ["./*"],
-      "@shared/*": ["./shared/*"],
       "@ui/*": ["./components/ui/*"],
       "@comp/*": ["./components/*"],
-      "@server/*": ["../server/*"]
-    }
+      "@server/*": ["../server/*"],
+    },
   },
   "references": [{ "path": "../server" }]
 }
