+++ code/mobile/update_shared.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+npm remove shared
+npm install git+https://github.com/FitnessApp-Clemens-and-Yanni/the-fitness-app-shared.git
