+++ code/server/package.json
@@ -21,6 +21,7 @@
     "express": "^5.1.0",
     "mongodb": "^6.15.0",
     "oauth-1.0a": "^2.2.6",
+    "shared": "file:../shared",
     "superjson": "^2.2.2",
     "zod": "^3.24.2"
   },
