+++ code/server/routers/food/food/getTargetNutritionalValuesPublicQuery.ts
@@ -1,0 +1,14 @@
+import { TARGET_NUTRITIONAL_VALUE_COLLECTION } from "@/data/meta/models.js";
+import { TargetNutritionalValue } from "@/data/meta/models.js";
+import { publicProcedure } from "@/trpc.js";
+import z from "zod";
+
+export const getTargetNutritionalValuesPublicQuery = publicProcedure
+  .input(z.object({ userId: z.string() }))
+  .query(async ({ ctx, input }) => {
+    const collection = ctx.db.collection<TargetNutritionalValue>(
+      TARGET_NUTRITIONAL_VALUE_COLLECTION,
+    );
+    const targetValues = await collection.findOne({ userId: input.userId });
+    return targetValues as TargetNutritionalValue;
+  });
