+++ code/server/routers/sport/snapshots/router-snapshots.ts
@@ -1,0 +1,6 @@
+import { createTRPCRouter } from "@/trpc.js";
+import { getExerciseDefaultsForWorkoutPublicQuery } from "./getExerciseDefaultsForWorkoutPublicQuery.js";
+
+export const SnapshotsRouter = createTRPCRouter({
+  getExerciseDefaultsForWorkout: getExerciseDefaultsForWorkoutPublicQuery,
+});
