+++ code/server/tsconfig.json
@@ -10,8 +10,7 @@
     "strict": true,
     "skipLibCheck": true,
     "paths": {
-      "@/*": ["./*"],
-      "@shared/*": ["./shared/*"]
+      "@/*": ["./*"]
     },
     "composite": true
   },
