+++ code/shared/tsconfig.json
@@ -1,16 +1,15 @@
 {
   "compilerOptions": {
     "target": "ESNext",
-    "module": "CommonJS",
-    "declaration": true,
-    "declarationMap": true,
-    "noEmit": true,
-    "rootDir": "src",
-    "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true,
-    "strict": true,
-    "skipLibCheck": true,
+    "module": "nodenext",
+    "moduleResolution": "nodenext",
     "sourceMap": true,
-    "composite": true
+    "composite": true,
+
+    "rootDir": "src",
+    "outDir": "build",
+
+    "declaration": true,
+    "strict": true,
   },
 }
